refactor(recognition): extract shared getCardReactionSummary helper
The groupBy + findMany + Map + REACTION_EMOJIS.map pattern was
duplicated in the dashboard detail page and public share page. Extract
into lib/interactions.ts so both pages call a single function.

The feed API's batch variant (multi-card, groups by cardId) is
architecturally different and stays as-is.
